LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-shadow-animations-expected.txt


FAIL Can animate standard property in shadow assert_equals: expected "0px" but got "784px"
FAIL Can animate typed custom property in shadow assert_equals: expected "40px" but got "0px"
FAIL Can animate untyped custom property in shadow assert_equals: expected "0px" but got "--from()"
